+++ components/codecs/include/ftl/codecs/packet.hpp
@@ -73,7 +73,7 @@ struct StreamPacket {
 	ftl::codecs::Channel channel;		// Actual channel of this current set of packets
 
 	inline int frameNumber() const { return (version >= 4) ? frame_number : streamID; }
-	inline int frameSetID() const { return (version >= 4) ? streamID : 0; }
+	inline size_t frameSetID() const { return (version >= 4) ? streamID : 0; }
 	inline int64_t localTimestamp() const { return timestamp + originClockDelta; }
 
 	int64_t originClockDelta;  // Not message packet / saved
